+++ src/main/java/net/geforcemods/securitycraft/SCContent.java
@@ -171,6 +171,7 @@ import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedRedstoneLampBlo
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedRootedDirtBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedRotatedCrystalQuartzPillar;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedRotatedPillarBlock;
+import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedScaffoldingBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedSlabBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedSnowyDirtBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedSoulSandBlock;
@@ -222,6 +223,7 @@ import net.geforcemods.securitycraft.items.LensItem;
 import net.geforcemods.securitycraft.items.MineRemoteAccessToolItem;
 import net.geforcemods.securitycraft.items.ModuleItem;
 import net.geforcemods.securitycraft.items.PortableTunePlayerItem;
+import net.geforcemods.securitycraft.items.ReinforcedScaffoldingBlockItem;
 import net.geforcemods.securitycraft.items.SCManualItem;
 import net.geforcemods.securitycraft.items.SecuritySeaBoatItem;
 import net.geforcemods.securitycraft.items.SentryItem;
@@ -2174,6 +2176,9 @@ public class SCContent {
 	@OwnableBE
 	@Reinforced
 	public static final RegistryObject<Block> REINFORCED_BLACK_GLAZED_TERRACOTTA = BLOCKS.register("reinforced_black_glazed_terracotta", () -> new ReinforcedGlazedTerracottaBlock(prop(DyeColor.BLACK.getMapColor()), Blocks.BLACK_GLAZED_TERRACOTTA));
+	@OwnableBE
+	@Reinforced(registerBlockItem = false)
+	public static final RegistryObject<Block> REINFORCED_SCAFFOLDING = BLOCKS.register("reinforced_scaffolding", () -> new ReinforcedScaffoldingBlock(reinforcedCopy(Blocks.SCAFFOLDING)));
 	@HasManualPage(PageGroup.REINFORCED)
 	@OwnableBE
 	@Reinforced
@@ -2620,6 +2625,8 @@ public class SCContent {
 	public static final RegistryObject<Item> SCANNER_DOOR_ITEM = ITEMS.register("scanner_door_item", () -> new DoubleHighBlockItem(SCANNER_DOOR.get(), itemProp()));
 	@HasManualPage
 	public static final RegistryObject<Item> SC_MANUAL = ITEMS.register("sc_manual", () -> new SCManualItem(itemProp(1)));
+	@HasManualPage(PageGroup.REINFORCED)
+	public static final RegistryObject<Item> REINFORCED_SCAFFOLDING_ITEM = ITEMS.register("reinforced_scaffolding", () -> new ReinforcedScaffoldingBlockItem(itemProp()));
 	@HasManualPage(PageGroup.SECRET_SIGNS)
 	public static final RegistryObject<Item> SECRET_OAK_SIGN_ITEM = ITEMS.register("secret_sign_item", () -> new SignItem(itemProp(16), SCContent.SECRET_OAK_SIGN.get(), SCContent.SECRET_OAK_WALL_SIGN.get()));
 	@HasManualPage(PageGroup.SECRET_SIGNS)